+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - core: update deps
 - templates: show technologies used in the report
+- templates: handle pluralization of 'vulnerabilities' word
 - domains: add global summary
 
 ## [0.4] - 2025-10-15